--- components/common/imageUploadWrapper/ImageUploadWrapper.type.ts
@@ -1,3 +1,0 @@
-export interface ImageUploadPropsType {
-    name: string;
-}
