--- components/common/modal/Modal.type.ts
@@ -1,4 +1,0 @@
-export interface ModalPropsType {
-    isShowing: boolean;
-    hide: (e: React.MouseEvent<HTMLDivElement, MouseEvent>) => void;
-}
