--- components/initialRegister/button/Button.type.ts
@@ -1,5 +1,0 @@
-export interface ButtonPropsType {
-    sort: string;
-    name?: string;
-    navigateToNext?: (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void;
-}
