+++ components/initialRegister/button/index.tsx
@@ -1,6 +1,11 @@
 import React from 'react';
 import * as S from './Button.style';
-import type { ButtonPropsType } from './Button.type';
+
+export interface ButtonPropsType {
+    sort: string;
+    name?: string;
+    navigateToNext?: (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void;
+}
 
 const Button: React.FC<ButtonPropsType> = ({ sort, name, navigateToNext }) => <S.StyledButton sort={sort} name={name} onClick={navigateToNext}>{name}</S.StyledButton>
