--- components/initialRegister/completeRegister/CompleteRegister.type.ts
@@ -1,1 +1,0 @@
-export {}
