+++ components/initialRegister/completeRegister/index.tsx
