--- components/initialRegister/setUserProfile/SetUserProfile.type.ts
@@ -1,1 +1,0 @@
-export {}
